requirements.txt
selenium
webdriver-manager
google-generativeai
google-genai
requests
python-dotenv
openai>=1.0.0